Se modifican varias cosas
